+++ telugu_panchangam/personal/slot_scorers.py
@@ -68,6 +68,8 @@ class _DayContext:
     allowed_nakshatras: frozenset
     prefer_nakshatras: frozenset
     allowed_tithi_numbers: frozenset
+    allowed_lagnas: frozenset
+    caution_lagna_solar: bool
     manual_checks: tuple[str, ...]